[Maps] allow for extensible sourceDescriptor typings (#74758)
* [Maps] allow for extensible sourceDescriptor typings

* revert changes to blended_vector_layer.test
